LFS-172: Fix Object Bug & Move into new Folder
