+++ modules/data-entry/src/main/frontend/src/questionnaireEditor/AnswerOptions.jsx
@@ -27,7 +27,7 @@ import {
   withStyles
 } from "@material-ui/core";
 
-import QuestionnaireStyle from './QuestionnaireStyle';
+import QuestionnaireStyle from '../questionnaire/QuestionnaireStyle';
 import uuid from 'uuid/v4';
 import CloseIcon from '@material-ui/icons/Close';
