+++ modules/data-entry/src/main/frontend/src/questionnaireEditor/EditDialog.jsx
@@ -31,7 +31,7 @@ import {
   withStyles
 } from "@material-ui/core";
 
-import QuestionnaireStyle from './QuestionnaireStyle';
+import QuestionnaireStyle from '../questionnaire/QuestionnaireStyle';
 import AnswerOptions from './AnswerOptions';
 import Fields from './Fields'
