+++ modules/data-entry/src/main/frontend/src/questionnaireEditor/NumberInput.jsx
@@ -17,40 +17,54 @@
 //  under the License.
 //
 
-import React from 'react';
+import React, { useState, useRef } from 'react';
 import PropTypes from 'prop-types';
-import { Grid, TextField, Typography, withStyles } from "@material-ui/core";
+import {
+  Grid,
+  TextField,
+  Typography,
+  withStyles
+} from "@material-ui/core";
 
-import QuestionnaireStyle from './QuestionnaireStyle';
+import QuestionnaireStyle from "../questionnaire/QuestionnaireStyle";
 import QuestionComponentManager from "./QuestionComponentManager";
 
-// Text Input field used by Edit dialog component
-let TextInput = (props) => {
-  let { objectKey, data } = props;
+// Number Input field used by Edit dialog component
 
+let NumberInput = (props) => {
+  let { objectKey, data, definition } = props;
   let formatString = (key) => {
     let formattedString = key.charAt(0).toUpperCase() + key.slice(1);
       return formattedString.split(/(?=[A-Z])/).join(' ');
   }
-
   return (
-    <Grid container alignItems='flex-end' spacing={2} key={objectKey}>
-      <Grid item xs={6}><Typography>{ formatString(objectKey)}</Typography></Grid>
+    <Grid container alignItems='flex-end' spacing={2} key={objectKey || ''}>
+      <Grid item xs={6}><Typography>{ formatString(objectKey) || ''}</Typography></Grid>
       <Grid item xs={6}>
-        <TextField name={objectKey} id={objectKey} defaultValue={data[objectKey] || ''} required={objectKey.includes('text')}/>
+        <TextField
+          name={objectKey || ''}
+          id={objectKey || ''}
+          defaultValue={data[objectKey] || ''}
+          type='number' 
+          placeholder={objectKey.includes('maxPerSubject') ? 'Unlimited' : ''}
+          min={objectKey.includes('maxPerSubject') ? 0 : ''}
+        />
       </Grid>
     </Grid>
   )
 }
 
-TextInput.propTypes = {
+NumberInput.propTypes = {
   objectKey: PropTypes.string.isRequired,
   data: PropTypes.object.isRequired
 };
 
-const StyledTextInput = withStyles(QuestionnaireStyle)(TextInput);
-export default StyledTextInput;
+const StyledNumberInput = withStyles(QuestionnaireStyle)(NumberInput);
+export default NumberInput;
 
 QuestionComponentManager.registerQuestionComponent((definition) => {
-  return [StyledTextInput, 0];
+  if (["long", "double", "decimal"].includes(definition)) {
+    return [StyledNumberInput, 50];
+  }
 });
+
