+++ modules/data-entry/src/main/frontend/src/questionnaireEditor/QuestionComponentManager.jsx
