+++ modules/data-entry/src/main/frontend/src/questionnaireEditor/Section.json
